Move torch_stat to tools
