+++ tests/tools/test_stats.py
@@ -1,7 +1,7 @@
 import os
 import shutil
 import torch
-from torch_ttnn import torch_stat
+from tools import torch_stat
 import unittest
 import json
 from torch.fx.passes.dialect.common.cse_pass import CSEPass
