+++ tools/run_transformers.py
@@ -178,7 +178,7 @@ if __name__ == "__main__":
     if args.backend == "torch_ttnn":
         import torch_ttnn
     elif args.backend == "torch_stat":
-        from torch_ttnn import torch_stat
+        from tools import torch_stat
 
     models = [
         TestModel(
